Fix bugs
