+++ .gitignore
@@ -21,5 +21,4 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-/src/classBased
-/src/functionBased
+/src/backUpCode
