+++ src/App.scss
@@ -161,3 +161,14 @@ a {
     line-height: 25px;
   }
 }
+
+.error {
+  text-align: center;
+  font-size: 16px;
+  color: red;
+  font-weight: bold;
+  background-color: rgba(red, 0.2);
+  padding: 8px;
+  border-radius: 4px;
+  visibility: hidden;
+}
