--- src/classBased/components/TodoContainer.js
@@ -1,110 +1,0 @@
-/* eslint-disable no-param-reassign */
-import React from 'react';
-import { v4 as uuidv4 } from 'uuid';
-import TodosList from './TodosList';
-import Header from './Header';
-import InputTodo from './InputTodo';
-
-class TodoContainer extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      todos: [],
-    };
-  }
-
-  componentDidMount() {
-    const temp = localStorage.getItem('todos');
-    const loadedTodos = JSON.parse(temp);
-    if (loadedTodos) {
-      this.setState({
-        todos: loadedTodos,
-      });
-    }
-  }
-
-  componentDidUpdate(prevProps, prevState) {
-    const { todos } = this.state;
-    if (prevState.todos !== todos) {
-      const temp = JSON.stringify(todos);
-      localStorage.setItem('todos', temp);
-    }
-  }
-
-  handleChange = (id) => {
-    this.setState((prevState) => ({
-      todos: prevState.todos.map((todo) => {
-        if (todo.id === id) {
-          return {
-            ...todo,
-            completed: !todo.completed,
-          };
-        }
-        return todo;
-      }),
-    }));
-    // this.setState((prevState) => ({
-    //   todos: prevState.todos.map((todo) => {
-    //     if (todo.id === id) {
-    //       return {
-    //         ...todo,
-    //         completed: !todo.completed,
-    //       };
-    //     }
-    //     return todo;
-    //   }),
-    // }));
-  };
-
-  delTodo = (id) => {
-    const { todos } = this.state;
-    this.setState({
-      todos: [
-        ...todos.filter((todo) => todo.id !== id),
-      ],
-    });
-  };
-
-  addTodoItem = (title) => {
-    const newTodo = {
-      id: uuidv4(),
-      title,
-      completed: false,
-    };
-    const { todos } = this.state;
-    this.setState({
-      todos: [...todos, newTodo],
-    });
-  };
-
-  setUpdate = (updatedTitle, id) => {
-    const { todos } = this.state;
-    this.setState({
-      todos: todos.map((todo) => {
-        if (todo.id === id) {
-          todo.title = updatedTitle;
-        }
-        return todo;
-      }),
-    });
-  }
-
-  render() {
-    const { todos } = this.state;
-    return (
-      <div className="container">
-        <div className="inner">
-          <Header />
-          <InputTodo addTodoProps={this.addTodoItem} />
-          <TodosList
-            todos={todos}
-            handleChangeProps={this.handleChange}
-            deleteTodoProps={this.delTodo}
-            setUpdate={this.setUpdate}
-          />
-        </div>
-      </div>
-    );
-  }
-}
-export default TodoContainer;
