--- src/functionBased/components/InputTodo.js
@@ -1,47 +1,0 @@
-import React, { useState } from 'react';
-import PropTypes from 'prop-types';
-import { FaPlusCircle } from 'react-icons/fa';
-
-const InputTodo = (props) => {
-  const [inputText, setInputText] = useState({
-    title: '',
-  });
-
-  const onChange = (e) => {
-    setInputText({
-      ...inputText,
-      [e.target.name]: e.target.value,
-    });
-  };
-
-  const handleSubmit = (e) => {
-    const { addTodoProps } = props;
-    e.preventDefault();
-    if (inputText.title.trim()) {
-      addTodoProps(inputText.title);
-      setInputText({
-        title: '',
-      });
-    } else {
-      alert('Please write item');
-    }
-  };
-
-  return (
-    <form onSubmit={handleSubmit} className="form-container">
-      <input
-        type="text"
-        className="input-text"
-        placeholder="Add todo..."
-        value={inputText.title}
-        name="title"
-        onChange={onChange}
-      />
-      <button type="button" className="input-submit" label="submit"><FaPlusCircle /></button>
-    </form>
-  );
-};
-InputTodo.propTypes = {
-  addTodoProps: PropTypes.func.isRequired,
-};
-export default InputTodo;
