--- src/functionBased/components/TodoItem.js
@@ -1,76 +1,0 @@
-import React, { useState, useEffect } from 'react';
-import PropTypes from 'prop-types';
-import { FaTrash } from 'react-icons/fa';
-import styles from './TodoItem.module.scss';
-
-const TodoItem = (props) => {
-  const [editing, setEditing] = useState(false);
-
-  const handleEditing = () => {
-    setEditing(true);
-  };
-
-  const handleUpdatedDone = (event) => {
-    if (event.key === 'Enter') {
-      setEditing(false);
-    }
-  };
-
-  const completedStyle = {
-    fontStyle: 'italic',
-    color: '#595959',
-    opacity: 0.4,
-    textDecoration: 'line-through',
-  };
-  const {
-    todo, handleChangeProps, deleteTodoProps, setUpdate,
-  } = props;
-  const { completed, id, title } = todo;
-
-  const viewMode = {};
-  const editMode = {};
-
-  if (editing) {
-    viewMode.display = 'none';
-  } else {
-    editMode.display = 'none';
-  }
-  useEffect(() => () => {
-    console.log('Cleaning up...');
-  }, []);
-  return (
-    <li className={styles.item}>
-      <div onDoubleClick={handleEditing} style={viewMode}>
-        <input
-          type="checkbox"
-          className={styles.checkbox}
-          checked={completed}
-          onChange={() => handleChangeProps(id)}
-        />
-        <button type="button" onClick={() => deleteTodoProps(id)} label="delete"><FaTrash style={{ color: 'orangered', fontSize: '16px' }} /></button>
-        <span style={completed ? completedStyle : null}>{title}</span>
-      </div>
-      <input
-        type="text"
-        style={editMode}
-        className={styles.textInput}
-        value={title}
-        onChange={(e) => {
-          setUpdate(e.target.value, id);
-        }}
-        onKeyDown={handleUpdatedDone}
-      />
-    </li>
-  );
-};
-TodoItem.propTypes = {
-  todo: PropTypes.shape({
-    id: PropTypes.string,
-    title: PropTypes.string,
-    completed: PropTypes.bool,
-  }).isRequired,
-  handleChangeProps: PropTypes.func.isRequired,
-  deleteTodoProps: PropTypes.func.isRequired,
-  setUpdate: PropTypes.func.isRequired,
-};
-export default TodoItem;
