--- src/functionBased/components/TodosList.js
@@ -1,35 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import TodoItem from './TodoItem';
-
-const TodosList = (props) => {
-  const {
-    todos, handleChangeProps, deleteTodoProps, setUpdate,
-  } = props;
-  return (
-    <ul>
-      {todos.map((todo) => (
-        <TodoItem
-          key={todo.id}
-          todo={todo}
-          handleChangeProps={handleChangeProps}
-          deleteTodoProps={deleteTodoProps}
-          setUpdate={setUpdate}
-        />
-      ))}
-    </ul>
-  );
-};
-TodosList.propTypes = {
-  todos: PropTypes.arrayOf(
-    PropTypes.shape({
-      id: PropTypes.string,
-      title: PropTypes.string,
-      completed: PropTypes.bool,
-    }),
-  ).isRequired,
-  handleChangeProps: PropTypes.func.isRequired,
-  deleteTodoProps: PropTypes.func.isRequired,
-  setUpdate: PropTypes.func.isRequired,
-};
-export default TodosList;
