+++ src/index.js
@@ -1,14 +1,14 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import { BrowserRouter as Router } from 'react-router-dom';
 import TodoContainer from './components/TodoContainer';
 import './App.scss';
-// eslint-disable-next-line
-ReactDOM.render(
+
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
   <React.StrictMode>
     <Router>
       <TodoContainer />
     </Router>
   </React.StrictMode>,
-  document.getElementById('root'),
 );
